requirements.txt
cox==0.1.post3
dill==0.3.9
faiss-gpu==1.7.1
joblib==1.4.2
numpy==1.24.4
pandas==2.0.3
requests==2.32.3
scikit-learn==1.3.2
tensorboardX==2.6.2.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1